Manager with Sponsor tab featuring table.
Started work on Manager.  Basically it just has three tabs right
now.  There is Sponsor, Spot, and Break.  Sponsor tab loads the
SponsorManager class.

SponsorManager features a table of all the sponsors in the
database.  The goal is that the user can select a sponsor from
the table and then edit or delete it by clicking one of the
buttons at the bottom.  Currently however the handler for the
buttons doesn't do anything with the actions.